apps/argus/pyproject.toml
[project]
name = "argus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "messages",
    "pandas>=2.3.3",
    "peewee>=3.18.2",
    "protobuf>=6.33.0",
    "psycopg2-binary>=2.9.11",
    "pyserial>=3.5",
]

[tool.uv.sources]
messages = { path = "../../common/messages/python", editable = true }
